a lot more refactoring to clean up the organisation
