+++ python/src/pyraphtory/api/__init__.py
@@ -1,6 +1,17 @@
-import pyraphtory.api.vertex
-import pyraphtory.api.input
-import pyraphtory.api.graph
-import pyraphtory.api.scala
-import pyraphtory.api.spouts
+"""
+Python wrappers for the raphtory api
+"""
+
+# Need to import all wrappers such that they are properly registered
+import pyraphtory.api.algorithm
 import pyraphtory.api.context
+import pyraphtory.api.formats
+import pyraphtory.api.graph
+import pyraphtory.api.graphstate
+import pyraphtory.api.input
+import pyraphtory.api.scala
+import pyraphtory.api.sinks
+import pyraphtory.api.sources
+import pyraphtory.api.spouts
+import pyraphtory.api.table
+import pyraphtory.api.vertex
