+++ python/src/pyraphtory/api/scala/__init__.py
@@ -1,2 +1,3 @@
+"""Python wrappers for core scala classes"""
 import pyraphtory.api.scala.collection
 import pyraphtory.api.scala.implicits
