+++ python/src/pyraphtory/api/scala/implicits/__init__.py
@@ -1,2 +1,3 @@
+"""Wrappers for scala implicit values"""
 import pyraphtory.api.scala.implicits.numeric
 import pyraphtory.api.scala.implicits.bounded
