+++ python/src/pyraphtory/api/sources.py
@@ -1,4 +1,4 @@
-from pyraphtory.api.interop import ScalaClassProxy
+from pyraphtory.interop import ScalaClassProxy
 
 class Source(ScalaClassProxy):
     _classname = "com.raphtory.api.input.Source"
