+++ python/src/pyraphtory/api/spouts.py
@@ -1,4 +1,4 @@
-from pyraphtory.api.interop import ScalaClassProxy
+from pyraphtory.interop import ScalaClassProxy
 
 
 """Spouts"""
