+++ python/src/pyraphtory/interop/__init__.py
@@ -1,0 +1,1 @@
+from ._interop import ScalaClassProxy, GenericScalaProxy, ScalaPackage, logger, register, to_jvm, to_python
