+++ python/src/pyraphtory/interop/_debug.py
