+++ python/src/pyraphtory/interop/_jpypeinterpreter.py
@@ -1,6 +1,5 @@
-import jpype.imports
-from jpype import JImplements, JOverride, JString, JObject, JClass
-from pyraphtory._jvm_interop.debug import enable_pydev_debug
+from jpype import JImplements, JOverride, JString, JObject, JClass, imports
+from pyraphtory.interop._debug import enable_pydev_debug
 from traceback import print_exc
 
 Interpreter = JClass("pemja.core.Interpreter")
